api/requirements.txt
Django==3.1.6
django-cors-headers==3.7.0
djangorestframework==3.12.2